Moved input-handling utils to UserInferface
